Automated-MCQ-Generator-Using-Langchain-GEMINI-API-main/requirements.txt
langchain[google-genai]
langchain
streamlit
python-dotenv
PyPDF2
